+++ CHANGELOG.md
@@ -16,12 +16,15 @@
 - `pcs status` works on remote nodes again (broken since pcs-0.10.4)
   ([rhbz#1830552])
 - Fixed inability to create colocation constraint from web ui ([rhbz#1832973])
+- Actions going through pcsd no longer time out after 30s (broken since
+  pcs-0.10.5) ([rhbz#1833506])
 
 [rhbz#1387358]: https://bugzilla.redhat.com/show_bug.cgi?id=1387358
 [rhbz#1782553]: https://bugzilla.redhat.com/show_bug.cgi?id=1782553
 [rhbz#1824206]: https://bugzilla.redhat.com/show_bug.cgi?id=1824206
 [rhbz#1830552]: https://bugzilla.redhat.com/show_bug.cgi?id=1830552
 [rhbz#1832973]: https://bugzilla.redhat.com/show_bug.cgi?id=1832973
+[rhbz#1833506]: https://bugzilla.redhat.com/show_bug.cgi?id=1833506
 
 
 ## [0.10.5] - 2020-03-18